Remove enums due to Rollup bug, add SubtitleFormat
