+++ docs/extras/modules/data_connection/document_loaders/integrations/web_loaders/assemblyai_audio_transcription.mdx
@@ -48,7 +48,10 @@ console.dir(docs, { depth: Infinity });
 You can also use the `AudioSubtitleLoader` to get `srt` or `vtt` subtitles as a document.
 
 ```typescript
-import { AudioSubtitleLoader } from "langchain/document_loaders/web/assemblyai";
+import { 
+  AudioSubtitleLoader, 
+  SubtitleFormat 
+} from "langchain/document_loaders/web/assemblyai";
 
 // You can also use a local file path and the loader will upload it to AssemblyAI for you
 const audioUrl = "https://storage.googleapis.com/aai-docs-samples/espn.m4a";
@@ -58,7 +61,7 @@ const loader = new AudioSubtitleLoader(
     audio_url: audioUrl
     // any other parameters as documented here: https://www.assemblyai.com/docs/API%20reference/transcript#create-a-transcript 
   },
-  "srt", // srt or vtt
+  SubtitleFormat.Srt, // srt or vtt
   {
     apiKey: "<ASSEMBLYAI_API_KEY>" // or set the `ASSEMBLYAI_API_KEY` env variable
   }
