+++ examples/src/document_loaders/assemblyai_audio_transcription.ts
@@ -2,7 +2,8 @@ import {
   AudioTranscriptLoader,
   AudioTranscriptParagraphsLoader,
   AudioTranscriptSentencesLoader,
-  AudioSubtitleLoader
+  AudioSubtitleLoader,
+  SubtitleFormat
 } from "langchain/document_loaders/web/assemblyai";
 
 // You can also use a local file path and the loader will upload it to AssemblyAI for you
@@ -11,23 +12,22 @@ const audioUrl = "https://storage.googleapis.com/aai-docs-samples/nbc.mp3";
 // use `AudioTranscriptParagraphsLoader` or `AudioTranscriptSentencesLoader` for splitting the transcript into paragraphs or sentences
 const transcriptLoader = new AudioTranscriptLoader(
   {
-    audio_url: audioUrl
+    audio_url: audioUrl,
   },
   {
-    apiKey: "<ASSEMBLYAI_API_KEY>" // or set the `ASSEMBLYAI_API_KEY` env variable
+    apiKey: "<ASSEMBLYAI_API_KEY>", // or set the `ASSEMBLYAI_API_KEY` env variable
   }
 );
 const transcriptDocs = await transcriptLoader.load();
 console.dir(transcriptDocs, { depth: Infinity });
 
-
 const subtitleLoader = new AudioSubtitleLoader(
   {
-    audio_url: audioUrl
+    audio_url: audioUrl,
   },
-  "srt", // srt or vtt
+  SubtitleFormat.Srt, // srt or vtt
   {
-    apiKey: "<ASSEMBLYAI_API_KEY>" // or set the `ASSEMBLYAI_API_KEY` env variable
+    apiKey: "<ASSEMBLYAI_API_KEY>", // or set the `ASSEMBLYAI_API_KEY` env variable
   }
 );
