--- langchain/src/types/assemblyai-types.d.ts
@@ -1,6 +1,0 @@
-export type assemblyAIOptions = {
-  /**
-   * The AssemblyAI API key. Alternatively, you can configure this as the `ASSEMBLYAI_API_KEY` environment variable.
-   */
-  apiKey: string;
-};
