+++ langchain/tsconfig.json
@@ -104,6 +104,7 @@
       "src/document.ts",
       "src/document_loaders/base.ts",
       "src/document_loaders/web/apify_dataset.ts",
+      "src/document_loaders/web/assemblyai.ts",
       "src/document_loaders/web/azure_blob_storage_container.ts",
       "src/document_loaders/web/azure_blob_storage_file.ts",
       "src/document_loaders/web/cheerio.ts",
